refactor: make the types module strictly type definitions
