+++ backend/internal/services/app_images_service.go
@@ -79,12 +79,12 @@ func (s *ApplicationImagesService) GetImageWithColor(name string, colorOverride 
 
 func (s *ApplicationImagesService) applyAccentColorToSVGInternal(svgData []byte, colorOverride string) []byte {
 	accentColor := ""
-	if settings.SafeAccentColor.MatchString(colorOverride) {
+	if safeAccentColor.MatchString(colorOverride) {
 		accentColor = colorOverride
 	} else if s.settingsService != nil {
 		if cfg := s.settingsService.GetSettingsConfig(); cfg != nil {
 			stored := cfg.AccentColor.Value
-			if stored != "" && stored != "default" && settings.SafeAccentColor.MatchString(stored) {
+			if stored != "" && stored != "default" && safeAccentColor.MatchString(stored) {
 				accentColor = stored
 			}
 		}
