+++ backend/internal/services/vulnerability_service.go
@@ -1614,7 +1614,7 @@ func (s *VulnerabilityService) execTrivyScanInContainer(ctx context.Context, con
 		return nil, errors.Errorf("trivy scan failed: %s", errMsg)
 	}
 
-	result := vulnerability.ConvertTrivyReportToScanResult(trivyReport, imageID, time.Now(), 0)
+	result := convertTrivyReportToScanResultInternal(trivyReport, imageID, time.Now(), 0)
 	result.ScannerVersion = scannerVersion
 
 	return result, nil
@@ -3421,7 +3421,7 @@ func (s *VulnerabilityService) runTrivyScan(ctx context.Context, trivyImage stri
 	}
 
 	// Convert Trivy report to our ScanResult format
-	result := vulnerability.ConvertTrivyReportToScanResult(trivyReport, imageID, time.Now(), duration)
+	result := convertTrivyReportToScanResultInternal(trivyReport, imageID, time.Now(), duration)
 	result.ScannerVersion = s.GetTrivyVersion(ctx)
 	totalVulnerabilities := 0
 	if result.Summary != nil {
@@ -3987,4 +3987,106 @@ func (s *VulnerabilityService) GetIgnoreRecordsForImage(ctx context.Context, env
 	}
 
 	return ignores, nil
+}
+
+func convertTrivyReportToScanResultInternal(report *vulnerability.TrivyReport, imageID string, scanTime time.Time, duration int64) *vulnerability.ScanResult {
+	result := &vulnerability.ScanResult{
+		ImageID:   imageID,
+		ImageName: report.ArtifactName,
+		ScanTime:  scanTime,
+		Status:    vulnerability.ScanStatusCompleted,
+		Duration:  duration,
+		Summary: &vulnerability.SeveritySummary{
+			Critical: 0,
+			High:     0,
+			Medium:   0,
+			Low:      0,
+			Unknown:  0,
+			Total:    0,
+		},
+		Vulnerabilities: []vulnerability.Vulnerability{},
+	}
+
+	for _, trivyResult := range report.Results {
+		for _, trivyVulnerability := range trivyResult.Vulnerabilities {
+			mappedVulnerability := convertTrivyVulnerabilityInternal(&trivyVulnerability)
+			result.Vulnerabilities = append(result.Vulnerabilities, mappedVulnerability)
+
+			switch mappedVulnerability.Severity {
+			case vulnerability.SeverityCritical:
+				result.Summary.Critical++
+			case vulnerability.SeverityHigh:
+				result.Summary.High++
+			case vulnerability.SeverityMedium:
+				result.Summary.Medium++
+			case vulnerability.SeverityLow:
+				result.Summary.Low++
+			case vulnerability.SeverityUnknown:
+				result.Summary.Unknown++
+			default:
+				result.Summary.Unknown++
+			}
+			result.Summary.Total++
+		}
+	}
+
+	return result
+}
+
+func convertTrivyVulnerabilityInternal(trivyVulnerability *vulnerability.TrivyVulnerability) vulnerability.Vulnerability {
+	mappedVulnerability := vulnerability.Vulnerability{
+		VulnerabilityID:  trivyVulnerability.VulnerabilityID,
+		PkgName:          trivyVulnerability.PkgName,
+		InstalledVersion: trivyVulnerability.InstalledVersion,
+		FixedVersion:     trivyVulnerability.FixedVersion,
+		Severity:         parseSeverityInternal(trivyVulnerability.Severity),
+		Title:            trivyVulnerability.Title,
+		Description:      trivyVulnerability.Description,
+		References:       trivyVulnerability.References,
+	}
+
+	if len(trivyVulnerability.CVSS) > 0 {
+		cvss := &vulnerability.CVSSInfo{}
+		for _, cvssData := range trivyVulnerability.CVSS {
+			if cvssData.V3Score > 0 {
+				cvss.V3Score = cvssData.V3Score
+				cvss.V3Vector = cvssData.V3Vector
+			}
+			if cvssData.V2Score > 0 {
+				cvss.V2Score = cvssData.V2Score
+				cvss.V2Vector = cvssData.V2Vector
+			}
+		}
+		if cvss.V2Score > 0 || cvss.V3Score > 0 {
+			mappedVulnerability.CVSS = cvss
+		}
+	}
+
+	if trivyVulnerability.PublishedDate != "" {
+		if parsedTime, err := time.Parse(time.RFC3339, trivyVulnerability.PublishedDate); err == nil {
+			mappedVulnerability.PublishedDate = &parsedTime
+		}
+	}
+	if trivyVulnerability.LastModifiedDate != "" {
+		if parsedTime, err := time.Parse(time.RFC3339, trivyVulnerability.LastModifiedDate); err == nil {
+			mappedVulnerability.LastModifiedDate = &parsedTime
+		}
+	}
+
+	return mappedVulnerability
+}
+
+func parseSeverityInternal(severity string) vulnerability.Severity {
+	switch severity {
+	case "CRITICAL":
+		return vulnerability.SeverityCritical
+	case "HIGH":
+		return vulnerability.SeverityHigh
+	case "MEDIUM":
+		return vulnerability.SeverityMedium
+	case "LOW":
+		return vulnerability.SeverityLow
+	default:
+		return vulnerability.SeverityUnknown
+	}
 }
