+++ types/image/image.go
@@ -2,13 +2,10 @@ package image
 
 import (
 	"encoding/json"
-	"math"
-	"strings"
 	"time"
 
 	"github.com/getarcaneapp/arcane/types/v2/containerregistry"
 	"github.com/getarcaneapp/arcane/types/v2/vulnerability"
-	"github.com/moby/moby/api/types/image"
 )
 
 type UpdateInfo struct {
@@ -230,32 +227,6 @@ type SearchResult struct {
 	Automated   bool   `json:"automated"`
 }
 
-// NewPruneReport creates a PruneReport from a Docker image prune report.
-// It extracts deleted and untagged image IDs from the Docker API response,
-// combining both types into a single list and converting space reclaimed to int64.
-func NewPruneReport(src image.PruneReport) PruneReport {
-	// Safely convert uint64 to int64, capping at MaxInt64 to prevent overflow
-	var spaceReclaimed int64
-	if src.SpaceReclaimed > uint64(math.MaxInt64) {
-		spaceReclaimed = math.MaxInt64
-	} else {
-		spaceReclaimed = int64(src.SpaceReclaimed)
-	}
-
-	out := PruneReport{
-		ImagesDeleted:  make([]string, 0, len(src.ImagesDeleted)),
-		SpaceReclaimed: spaceReclaimed,
-	}
-	for _, d := range src.ImagesDeleted {
-		if d.Deleted != "" {
-			out.ImagesDeleted = append(out.ImagesDeleted, d.Deleted)
-		} else if d.Untagged != "" {
-			out.ImagesDeleted = append(out.ImagesDeleted, d.Untagged)
-		}
-	}
-	return out
-}
-
 type UsageCounts struct {
 	// Inuse is the number of images currently in use.
 	//
@@ -411,99 +382,4 @@ type PullOptions struct {
 	//
 	// Required: false
 	Credentials []containerregistry.Credential `json:"credentials,omitempty"`
-}
-
-// GetFullImageName returns the image name with tag.
-func (p PullOptions) GetFullImageName() string {
-	if p.Tag != "" && p.Tag != "latest" {
-		return p.ImageName + ":" + p.Tag
-	}
-	if p.Tag == "latest" && !strings.Contains(p.ImageName, ":") {
-		return p.ImageName + ":latest"
-	}
-	return p.ImageName
-}
-
-// GetCredentials returns credentials from either the Auth or Credentials field.
-func (p PullOptions) GetCredentials() []containerregistry.Credential {
-	if len(p.Credentials) > 0 {
-		return p.Credentials
-	}
-	if p.Auth != nil {
-		return []containerregistry.Credential{*p.Auth}
-	}
-	return nil
-}
-
-// NewDetailSummary creates a DetailSummary from a Docker image inspect response.
-// It converts the Docker API types to the application's DetailSummary type,
-// handling nested structs and converting exposed ports from Docker's nat.PortSet
-// to string keys. The descriptor is derived from the first repo digest if available.
-func NewDetailSummary(src *image.InspectResponse) DetailSummary {
-	var out DetailSummary
-	if src == nil {
-		return out
-	}
-
-	out.ID = src.ID
-	out.RepoTags = append(out.RepoTags, src.RepoTags...)
-	out.RepoDigests = append(out.RepoDigests, src.RepoDigests...)
-	out.Comment = src.Comment
-	out.Created = src.Created
-	out.Author = src.Author
-
-	if src.Config != nil {
-		if len(src.Config.ExposedPorts) > 0 {
-			out.Config.ExposedPorts = make(map[string]struct{}, len(src.Config.ExposedPorts))
-			for p := range src.Config.ExposedPorts {
-				out.Config.ExposedPorts[p] = struct{}{}
-			}
-		}
-		if len(src.Config.Env) > 0 {
-			out.Config.Env = append(out.Config.Env, src.Config.Env...)
-		}
-		if len(src.Config.Cmd) > 0 {
-			out.Config.Cmd = append(out.Config.Cmd, src.Config.Cmd...)
-		}
-		if len(src.Config.Volumes) > 0 {
-			out.Config.Volumes = make(map[string]struct{}, len(src.Config.Volumes))
-			for v := range src.Config.Volumes {
-				out.Config.Volumes[v] = struct{}{}
-			}
-		}
-		out.Config.WorkingDir = src.Config.WorkingDir
-		out.Config.ArgsEscaped = src.Config.ArgsEscaped //nolint:staticcheck,nolintlint // Mirror Docker inspect data; deprecated only for new image builders.
-	}
-
-	out.Architecture = src.Architecture
-	out.Os = src.Os
-	out.Size = src.Size
-
-	if src.GraphDriver != nil {
-		out.GraphDriver.Name = src.GraphDriver.Name
-		if src.GraphDriver.Data != nil {
-			out.GraphDriver.Data = src.GraphDriver.Data
-		}
-	}
-
-	out.RootFs.Type = src.RootFS.Type
-	if len(src.RootFS.Layers) > 0 {
-		out.RootFs.Layers = append(out.RootFs.Layers, src.RootFS.Layers...)
-	}
-
-	if !src.Metadata.LastTagTime.IsZero() {
-		out.Metadata.LastTagTime = src.Metadata.LastTagTime.Format(time.RFC3339Nano)
-	}
-
-	// Best-effort descriptor from first digest
-	out.Descriptor.MediaType = "application/vnd.oci.image.index.v1+json"
-	out.Descriptor.Size = src.Size
-	if len(src.RepoDigests) > 0 {
-		parts := strings.SplitN(src.RepoDigests[0], "@", 2)
-		if len(parts) == 2 {
-			out.Descriptor.Digest = parts[1]
-		}
-	}
-
-	return out
 }
