+++ types/swarm/service.go
@@ -297,147 +297,3 @@ type ServiceScaleRequest struct {
 	// Required: true
 	Replicas uint64 `json:"replicas"`
 }
-
-// NewServiceSummary converts a Docker swarm service into the API-facing ServiceSummary shape.
-//
-// It derives the service mode, replica counts, running task counts, published
-// ports, stack namespace label, attached network names, and container mounts.
-// Network IDs are resolved through networkNameByID when possible, then fall
-// back to the attachment alias or raw target value. A nil nodeNames slice is
-// normalized to an empty slice for stable JSON output.
-//
-// service is the Docker swarm service to summarize.
-// nodeNames lists the node hostnames currently running tasks for the service.
-// networkNameByID maps attached network IDs to human-readable names.
-//
-// Returns a ServiceSummary populated from service and the supplied enrichment data.
-func NewServiceSummary(service swarm.Service, nodeNames []string, networkNameByID map[string]string) ServiceSummary {
-	spec := service.Spec
-
-	mode := "unknown"
-	replicas := uint64(0)
-	runningReplicas := uint64(0)
-	switch {
-	case spec.Mode.Replicated != nil:
-		mode = "replicated"
-		if spec.Mode.Replicated.Replicas != nil {
-			replicas = *spec.Mode.Replicated.Replicas
-		}
-	case spec.Mode.Global != nil:
-		mode = "global"
-		if service.ServiceStatus != nil {
-			replicas = service.ServiceStatus.DesiredTasks
-		}
-	case spec.Mode.ReplicatedJob != nil:
-		mode = "replicated-job"
-		switch {
-		case spec.Mode.ReplicatedJob.TotalCompletions != nil:
-			replicas = *spec.Mode.ReplicatedJob.TotalCompletions
-		case spec.Mode.ReplicatedJob.MaxConcurrent != nil:
-			replicas = *spec.Mode.ReplicatedJob.MaxConcurrent
-		default:
-			replicas = 1
-		}
-	case spec.Mode.GlobalJob != nil:
-		mode = "global-job"
-		if service.ServiceStatus != nil {
-			replicas = service.ServiceStatus.DesiredTasks
-		}
-	}
-
-	if service.ServiceStatus != nil {
-		runningReplicas = service.ServiceStatus.RunningTasks
-	}
-
-	image := ""
-	if spec.TaskTemplate.ContainerSpec != nil {
-		image = spec.TaskTemplate.ContainerSpec.Image
-	}
-
-	portSpecs := service.Endpoint.Spec.Ports
-	if len(portSpecs) == 0 {
-		portSpecs = service.Endpoint.Ports
-	}
-	ports := make([]ServicePort, 0, len(portSpecs))
-	for _, port := range portSpecs {
-		ports = append(ports, ServicePort{
-			Protocol:      string(port.Protocol),
-			TargetPort:    port.TargetPort,
-			PublishedPort: port.PublishedPort,
-			PublishMode:   string(port.PublishMode),
-		})
-	}
-
-	stackName := ""
-	if spec.Labels != nil {
-		stackName = spec.Labels[StackNamespaceLabel]
-	}
-
-	// Extract networks from task template, resolving IDs to names
-	networkConfigs := spec.TaskTemplate.Networks
-	networks := make([]string, 0, len(networkConfigs))
-	for _, n := range networkConfigs {
-		if name, ok := networkNameByID[n.Target]; ok {
-			networks = append(networks, name)
-		} else if len(n.Aliases) > 0 {
-			networks = append(networks, n.Aliases[0])
-		} else {
-			networks = append(networks, n.Target)
-		}
-	}
-
-	// Extract mounts from container spec
-	mounts := make([]ServiceMount, 0)
-	if spec.TaskTemplate.ContainerSpec != nil {
-		for _, m := range spec.TaskTemplate.ContainerSpec.Mounts {
-			mounts = append(mounts, ServiceMount{
-				Type:     string(m.Type),
-				Source:   m.Source,
-				Target:   m.Target,
-				ReadOnly: m.ReadOnly,
-			})
-		}
-	}
-
-	if nodeNames == nil {
-		nodeNames = []string{}
-	}
-
-	return ServiceSummary{
-		ID:              service.ID,
-		Name:            spec.Name,
-		Image:           image,
-		Mode:            mode,
-		Replicas:        replicas,
-		RunningReplicas: runningReplicas,
-		Ports:           ports,
-		CreatedAt:       service.CreatedAt,
-		UpdatedAt:       service.UpdatedAt,
-		Labels:          spec.Labels,
-		StackName:       stackName,
-		Nodes:           nodeNames,
-		Networks:        networks,
-		Mounts:          mounts,
-	}
-}
-
-// NewServiceInspect converts a Docker swarm service into the API-facing ServiceInspect shape.
-//
-// It copies the core inspection fields directly from the Docker SDK type.
-// Callers can enrich the returned value with node, network, or mount details
-// after construction when needed.
-//
-// service is the Docker swarm service to convert.
-//
-// Returns the base inspection payload for service.
-func NewServiceInspect(service swarm.Service) ServiceInspect {
-	return ServiceInspect{
-		ID:           service.ID,
-		Version:      service.Version,
-		CreatedAt:    service.CreatedAt,
-		UpdatedAt:    service.UpdatedAt,
-		Spec:         service.Spec,
-		Endpoint:     service.Endpoint,
-		UpdateStatus: service.UpdateStatus,
-	}
-}
