+++ types/vulnerability/vulnerability.go
@@ -365,113 +365,6 @@ type TrivyCVSS struct {
 	V3Score  float64 `json:"V3Score"`
 }
 
-// ConvertTrivyReportToScanResult converts a TrivyReport to a ScanResult
-func ConvertTrivyReportToScanResult(report *TrivyReport, imageID string, scanTime time.Time, duration int64) *ScanResult {
-	result := &ScanResult{
-		ImageID:   imageID,
-		ImageName: report.ArtifactName,
-		ScanTime:  scanTime,
-		Status:    ScanStatusCompleted,
-		Duration:  duration,
-		Summary: &SeveritySummary{
-			Critical: 0,
-			High:     0,
-			Medium:   0,
-			Low:      0,
-			Unknown:  0,
-			Total:    0,
-		},
-		Vulnerabilities: []Vulnerability{},
-	}
-
-	// Collect all vulnerabilities from all results
-	for _, trivyResult := range report.Results {
-		for _, tv := range trivyResult.Vulnerabilities {
-			vuln := convertTrivyVulnerability(&tv)
-			result.Vulnerabilities = append(result.Vulnerabilities, vuln)
-
-			// Update severity summary
-			switch vuln.Severity {
-			case SeverityCritical:
-				result.Summary.Critical++
-			case SeverityHigh:
-				result.Summary.High++
-			case SeverityMedium:
-				result.Summary.Medium++
-			case SeverityLow:
-				result.Summary.Low++
-			case SeverityUnknown:
-				result.Summary.Unknown++
-			default:
-				result.Summary.Unknown++
-			}
-			result.Summary.Total++
-		}
-	}
-
-	return result
-}
-
-func convertTrivyVulnerability(tv *TrivyVulnerability) Vulnerability {
-	vuln := Vulnerability{
-		VulnerabilityID:  tv.VulnerabilityID,
-		PkgName:          tv.PkgName,
-		InstalledVersion: tv.InstalledVersion,
-		FixedVersion:     tv.FixedVersion,
-		Severity:         parseSeverity(tv.Severity),
-		Title:            tv.Title,
-		Description:      tv.Description,
-		References:       tv.References,
-	}
-
-	// Convert CVSS scores
-	if len(tv.CVSS) > 0 {
-		cvss := &CVSSInfo{}
-		for _, cvssData := range tv.CVSS {
-			if cvssData.V3Score > 0 {
-				cvss.V3Score = cvssData.V3Score
-				cvss.V3Vector = cvssData.V3Vector
-			}
-			if cvssData.V2Score > 0 {
-				cvss.V2Score = cvssData.V2Score
-				cvss.V2Vector = cvssData.V2Vector
-			}
-		}
-		if cvss.V2Score > 0 || cvss.V3Score > 0 {
-			vuln.CVSS = cvss
-		}
-	}
-
-	// Parse dates
-	if tv.PublishedDate != "" {
-		if t, err := time.Parse(time.RFC3339, tv.PublishedDate); err == nil {
-			vuln.PublishedDate = &t
-		}
-	}
-	if tv.LastModifiedDate != "" {
-		if t, err := time.Parse(time.RFC3339, tv.LastModifiedDate); err == nil {
-			vuln.LastModifiedDate = &t
-		}
-	}
-
-	return vuln
-}
-
-func parseSeverity(s string) Severity {
-	switch s {
-	case "CRITICAL":
-		return SeverityCritical
-	case "HIGH":
-		return SeverityHigh
-	case "MEDIUM":
-		return SeverityMedium
-	case "LOW":
-		return SeverityLow
-	default:
-		return SeverityUnknown
-	}
-}
-
 // IgnorePayload represents the request to ignore a vulnerability
 type IgnorePayload struct {
 	// ImageID is the Docker image ID
